feat: redux to jotai, multicall state (#328)
* fix: multicall jotai

* fix: pools listing issue when aply jotai rerender issue

* fix: remove reducer action
